+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Quick token actions for Items without images caused visual glitches
 - API calls for dice rolls properly return and resolve to their result(s)
 - Removes items from containers when dragged out to same actor (prevents duplication) (thanks mkahvi!)
+- Initiative wasn't deferring to "CONFIG.Combat.initiative.formula" setting as per core
 
 ### Changelog